+++ config/settings/base.py
@@ -586,8 +586,6 @@ PROJECTROLES_SIDEBAR_ICON_SIZE = env.int('PROJECTROLES_SIDEBAR_ICON_SIZE', 36)
 PROJECTROLES_HIDE_PROJECT_APPS = env.list(
     'PROJECTROLES_HIDE_PROJECT_APPS', None, []
 )
-# NOTE: This setting has been deprecated and will be removed in v0.14
-PROJECTROLES_HIDE_APP_LINKS = env.list('PROJECTROLES_HIDE_APP_LINKS', None, [])
 
 # Set limit for delegate roles per project (if 0, no limit is applied)
 PROJECTROLES_DELEGATE_LIMIT = env.int('PROJECTROLES_DELEGATE_LIMIT', 1)